parameterizing the scripts

+++ dockerbuild.sh
@@ -1,6 +1,11 @@
 #!/bin/bash -e
 export PATH=/usr/local/bin:/path/to/node:/path/to/node_bin:/path/to/phantomjs:/path/to/jscoverage:$PATH;
 
+if [ "$1" == "" ]; then
+	echo "Usage: ./dockerbuild.sh <docker username>"
+	exit
+fi
+
 echo Cleaning...
 rm -rf ./dist
 
@@ -15,6 +20,6 @@ cd dist
 npm install --production
 
 echo Building docker image
-docker build -t agirmar/tictactoe .
+docker build -t "$1"/tictactoe .
 
 echo "Done"
